ansible-drift: ignore errors while installing roles and collections
This should ideally improve error handling in regards to connection
issues, like "The read operation timed out", which seem to happen more
frequently.

Obviously, that's not a proper solution, as calls which throw transient
errors should probably be retried, instead of being ignored and skipped
over, but it might help as a quick fix.